+++ NEWS.md
@@ -4,6 +4,7 @@
 - fix bug in `aqStats` if only a few lines of data
 - allow `statistic` to equal "mean", "median" or "frequency" in `scatterPlot`, when `method = "level"`
 - change smoothing in `scatterPlot` when `method = "level"` to use tensor interaction to allow for better smoothing when x and y are on different scales.
+- warn when >10% negative data detected in `polarPlot` and suggest setting `force.postive = FALSE`
 
 # openair 2.0-0