+++ NEWS.md
@@ -5,6 +5,7 @@
 - allow `statistic` to equal "mean", "median" or "frequency" in `scatterPlot`, when `method = "level"`
 - change smoothing in `scatterPlot` when `method = "level"` to use tensor interaction to allow for better smoothing when x and y are on different scales.
 - warn when >10% negative data detected in `polarPlot` and suggest setting `force.postive = FALSE`
+- make `maps` Suggests rather than Depends
 
 # openair 2.0-0