+++ NEWS.md
@@ -7,6 +7,7 @@
 - warn when >10% negative data detected in `polarPlot` and suggest setting `force.postive = FALSE`
 - make `maps` Suggests rather than Depends
 - refine check on whether bias correction is needed in `windRose`
+- still plot data when <6 points in `TheilSen` and when no trend information is given
 
 # openair 2.0-0