+++ NEWS.md
@@ -1,4 +1,6 @@
-# openair 2.0-99
+# openair 2.1-0
+
+## Main changes
 
 - add `angle` option to `percentileRose` to allow wind direction averages for sectors >10 degrees.
 - fix bug in `aqStats` if only a few lines of data